rebase with accrued rewards

+++ src/interfaces/IEtherFiOracle.sol
@@ -8,7 +8,7 @@ interface IEtherFiOracle {
         uint32 refSlotTo;
         uint32 refBlockFrom;
         uint32 refBlockTo;
-        int256 accruedRewards;
+        int128 accruedRewards;
         uint256[] validatorsToApprove;
         uint256[] liquidityPoolValidatorsToExit;
         uint256[] exitedValidators;
